requirements.txt
face_recognition==1.3.0
icrawler==0.6.10
matplotlib==3.8.4
numpy==2.2
Pillow==11.3.0
scikit_learn==1.7.0
seaborn==0.13.2
streamlit==1.46.1
torch==2.7.1
torchvision==0.22.1
